added a simple article activity with a list fragment for tablet landscape mode

+++ real-estate/src/main/java/com/zawadz88/realestate/RealEstateApplication.java
@@ -89,7 +89,7 @@ public class RealEstateApplication extends Application implements AsyncTaskListe
 		int result = -1;
 		Integer currentNumber = mArticleEssentialCurrentPageNumbersByCategory.get(categoryName);
 		if (currentNumber != null) {
-			result = currentNumber.intValue();
+			result = currentNumber;
 		}
 		return result;
 	}
